requirements.txt
aiogram~=3.22.0
Telethon~=1.41.2
python-dotenv~=1.2.1
aiosqlite~=0.21.0
aiohttp~=3.12.15